+++ src/main/java/gay/sylv/weird_wares/impl/util/Codecs.java
@@ -7,11 +7,13 @@
  */
 package gay.sylv.weird_wares.impl.util;
 
-import com.mojang.serialization.Codec;
+import com.mojang.datafixers.util.Pair;
+import com.mojang.serialization.*;
 import net.minecraft.core.BlockPos;
 
 import java.util.ArrayList;
 import java.util.HashSet;
+import java.util.List;
 import java.util.Set;
 
 @org.jetbrains.annotations.ApiStatus.Internal
@@ -19,4 +21,22 @@ public final class Codecs {
 	public static final Codec<Set<BlockPos>> POS_LIST = Codec.list(BlockPos.CODEC).xmap(HashSet::new, ArrayList::new);
 	
 	private Codecs() {}
+	
+	public static <T> Codec<Set<T>> set(Codec<T> elementCodec) {
+		Codec<List<T>> listCodec = Codec.list(elementCodec);
+		return Codec.of(
+				new Encoder<>() {
+					@Override
+					public <U> DataResult<U> encode(Set<T> input, DynamicOps<U> ops, U prefix) {
+						return listCodec.encode(List.copyOf(input), ops, prefix);
+					}
+				},
+				new Decoder<>() {
+					@Override
+					public <U> DataResult<Pair<Set<T>, U>> decode(DynamicOps<U> ops, U input) {
+						return listCodec.decode(ops, input).map((a) -> Pair.of(new HashSet<>(a.getFirst()), a.getSecond()));
+					}
+				}
+		);
+	}
 }