+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 # Weird Wares `0.3.0+1.21`
+## Added
+- Sculk remotes can shift right-click if the player holds shift and left-clicks a block.
+
 ## Changed
 - Disable glint rendering & remove from creative tab.